requirements.txt
fastapi
uvicorn
llama-cpp-python
sentence-transformers
faiss-cpu
PyMuPDF
python-multipart
